[API-91] Extract bootstrap.ts and reduce index.ts to a thin entrypoint.
