+++ api/.test.ts
@@ -2,7 +2,10 @@ import { describe, it, expect } from 'bun:test';
 import type { ActivityDto, ActivityListParams, ActivityListResult } from '@/activity';
 import { encodeCursor } from '@/util/cursor';
 import type { AlertDto } from '@/alerts';
-import { buildApp, gracefulShutdown, readExpoAccessToken, wrapScheduleWithReplan, wrapSystemWithReplan, type ScheduleApi, type SystemApi } from '@/index';
+import { buildApp } from '@/build-app';
+import { gracefulShutdown, readExpoAccessToken } from '@/bootstrap';
+import { wrapScheduleWithReplan, type ScheduleApi } from '@/routes/schedule';
+import { wrapSystemWithReplan, type SystemApi } from '@/routes/system';
 import type { TonightDto } from '@/models/tonight';
 import type { ScheduleListItem } from '@/service/schedules-list';
 import type { DaemonControl, DaemonStatus } from '@/service/daemon';
